+++ src/elements/TableTextElement.js
@@ -238,7 +238,7 @@ export default class TableTextElement extends TextElement {
             'cs_horizontalAlignment', 'cs_verticalAlignment', 'cs_textColor', 'cs_backgroundColor',
             'cs_font', 'cs_fontSize', 'cs_lineSpacing',
             'cs_paddingLeft', 'cs_paddingTop', 'cs_paddingRight', 'cs_paddingBottom',
-            'spreadsheet_textWrap'
+            'spreadsheet_type', 'spreadsheet_pattern', 'spreadsheet_textWrap'
         ];
         let tableBandObj = this.rb.getDataObject(this.parentId);
         if (tableBandObj !== null && tableBandObj.getValue('bandType') === Band.bandType.header) {